Add -file-list-shared-libraries MI command
This change adds the MI equivalent for the "info sharedlibrary"
command. The command was already partially documented but ignored as
it was not implemented. The new MI command works similarly to the CLI
command, taking an optional regular expression as an argument and
outputting the library information.

I included a test for the new command in mi-solib.exp.

gdb/doc/ChangeLog:

        * gdb.texinfo (gdb/mi Symbol Query Commands): Document new MI
        command file-list-shared-libraries

gdb/ChangeLog:

        * mi/mi-cmd-file.c (mi_cmd_file_list_shared_libraries):
        New function definition.
        * mi/mi-cmds.c (mi_cmds): Add -file-list-shared-libraries command.
        * mi/mi-cmds.h (mi_cmd_file_list_shared_libraries):
        New function declaration.
        * solib.c (info_sharedlibrary_command): Replace for loop with
        ALL_SO_LIBS macro
        * solib.h (update_solib_list): New function declaration.
        * solist.h (ALL_SO_LIBS): New macro.

gdb/testsuite/ChangeLog:

        * gdb.mi/mi-solib.exp (test_file_list_shared_libraries):
        New procedure.

Signed-off-by: Marc-Andre Laperle <malaperle@gmail.com>