Add a better diagnostic message in mi_gdb_test
When using mi_gdb_test, if it fails because of the presence of
unexpected output, the error message is only the message passed as
the argument with no indication that there was an unexpected output.
This change adds an additional text to the failure message to
indicate that there was an unexpected output.

gdb/testsuite/ChangeLog:

        * lib/mi-support.exp (mi_gdb_test): Add additional message
        for unexpected output.

Signed-off-by: Marc-Andre Laperle <malaperle@gmail.com>